// frontend/src/openapi/models/Car.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Car Rental API
 * Backend API for Car Rental Application
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Car
 */
export interface Car {
    /**
     * 
     * @type {number}
     * @memberof Car
     */
    id: number;
    /**
     * Car name/brand
     * @type {string}
     * @memberof Car
     */
    name: string;
    /**
     * Car model
     * @type {string}
     * @memberof Car
     */
    model: string;
    /**
     * Daily rental price
     * @type {string}
     * @memberof Car
     */
    pricePerDay: string;
    /**
     * Whether the car is available for booking
     * @type {boolean}
     * @memberof Car
     */
    isAvailable: boolean;
    /**
     * 
     * @type {number}
     * @memberof Car
     */
    latitude?: number | null;
    /**
     * 
     * @type {number}
     * @memberof Car
     */
    longitude?: number | null;
}

/**
 * Check if a given object implements the Car interface.
 */
export function instanceOfCar(value: object): value is Car {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('model' in value) || value['model'] === undefined) return false;
    if (!('pricePerDay' in value) || value['pricePerDay'] === undefined) return false;
    if (!('isAvailable' in value) || value['isAvailable'] === undefined) return false;
    return true;
}

export function CarFromJSON(json: any): Car {
    return CarFromJSONTyped(json, false);
}

export function CarFromJSONTyped(json: any, ignoreDiscriminator: boolean): Car {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'model': json['model'],
        'pricePerDay': json['price_per_day'],
        'isAvailable': json['is_available'],
        'latitude': json['latitude'] == null ? undefined : json['latitude'],
        'longitude': json['longitude'] == null ? undefined : json['longitude'],
    };
}

export function CarToJSON(json: any): Car {
    return CarToJSONTyped(json, false);
}

export function CarToJSONTyped(value?: Car | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'model': value['model'],
        'price_per_day': value['pricePerDay'],
        'is_available': value['isAvailable'],
        'latitude': value['latitude'],
        'longitude': value['longitude'],
    };
}